ref(builders): eliminate TRAINING tag, fix ExprKind priority, clean up registry
- Remove FittedPipelineTagKey.TRAINING; features/target already in ALL_STEPS
- Add FittedPipeline.from_tag_node for unambiguous recovery from specific tag
- Rewrite FittedPipeline.from_expr as outermost-pipeline convenience
- Store training_hash as inert provenance in ALL_STEPS metadata
- Fix ExprKind priority: Composed > ExprBuilder (preserves existing behavior)
- Add is_builder flag to ExprMetadata for orthogonal builder detection
- Add ExprBuilder to catalog run match arm
- Replace BuiltinTagKey enum with derived frozenset from _register_builtins
- Replace dict-truthiness init guard with boolean _initialized flag
- Protect builtins in _discover_from_tagged entry-point loading
- Add _reset_registry() for test isolation
- Split sklearn-dependent tests into test_builders_ml.py
- Guard BSL handler for missing boring-semantic-layer dependency
- Document _resolve_builder_from_tag traversal order
- Fix semantic_builder_example.py to use xo.memtable
- Add ADR-0005 for ExprBuilder/from_tagged registry design
