+++ python/xorq/expr/ml/enums.py
@@ -28,7 +28,6 @@ class FittedStepTagKey(StrEnum):
 
 class FittedPipelineTagKey(StrEnum):
     ALL_STEPS = "FittedPipeline-all_steps"
-    TRAINING = "FittedPipeline-training"
     TRANSFORM = "FittedPipeline-transform"
     PREDICT = "FittedPipeline-predict"
     PREDICT_PROBA = "FittedPipeline-predict_proba"
